profiling, new and reworked tests, multiple fixes.
	modified:   Makefile.common
	modified:   include/alu_box.h
	modified:   include/csx.h
	modified:   include/csx_test.h
	modified:   include/csx_test_arm_inst.h
	modified:   include/soc.h
	modified:   include/soc_core_cp15.h
	modified:   source/csx.c
	modified:   source/csx_test.c
	modified:   source/csx_test_arm.c
	modified:   source/csx_test_arm_inst.c
	modified:   source/csx_test_thumb.c
	modified:   source/soc.c
	modified:   source/soc_core.c
	modified:   source/soc_core_arm.c
	modified:   source/soc_core_decode.c
	modified:   source/soc_core_strings.c
	modified:   source/soc_core_trace_arm.c
	modified:   source/soc_mmio.c

+++ source/soc_core_decode.c
@@ -69,7 +69,7 @@ static void _soc_core_arm_shifter_operation_asr(soc_core_p core)
 
 static void _soc_core_arm_shifter_operation_lsl(soc_core_p core)
 {
-	uint cout = 0;
+	unsigned long int cout = 0;
 	
 	if(vR(S))
 		vR(SOP_V) = _lsl_c(vR(M), vR(S), &cout);
@@ -83,7 +83,7 @@ static void _soc_core_arm_shifter_operation_lsl(soc_core_p core)
 
 static void _soc_core_arm_shifter_operation_lsr(soc_core_p core)
 {
-	uint cout = 0;
+	unsigned long int cout = 0;
 
 	if(vR(S))
 		vR(SOP_V) = _lsr_c(vR(M), vR(S), &cout);
